Update plugin-with-data-access-4.30.md
Works fine when the user views a product and there are no or products that have been previously viewed . how ever once any user views the product details for a second time and exception is thrown indicating that it is unable to insert the log entry into the database table due to duplicate primary keys.. this is due to the entry being inserted will have an id of 0. The fix to to adjust the primary key to be a identity column in order to accommodate the insertion of multiple eateries each with their own unique identifier(unique primary key) for each log entry add to the table.